Remove unused is_reference_or_pointer_diff.
The function is_reference_or_pointer_diff was added in commit

85929105 Fix redundancy marking for change of types used directly

and was updated to peel typedefs as a first step in

ef9d20c9 Fix redundancy detection through fn ptr and typedef paths

which, however, also made it obsolete.

This commit removes the function's declaration and definition.
There are no functional changes.

	* include/abg-comparison.h (is_reference_or_pointer_diff):
	Drop function declaration.
	* src/abg-comparison.cc (is_reference_or_pointer_diff): Drop
	function definition.

Signed-off-by: Giuliano Procida <gprocida@google.com>