Use flat representation to canonicalize anonymous classes and unions
During the canonicalization of a type T, the algorithm uses the
internal pretty representation of T to limit the number of types to
compare T to.  That internal pretty representation is based on the
type name.

For anonymous types, the type name is not unique; it's constructed
just for internal purposes.  So using that in the pretty
representation might negatively impact the accuracy of the
canonicalization; it might make it so that two types might wrongly be
considered canonicaly different.

To fix that, this change makes the internal pretty representation of
anonymous classes (and unions) use their flat representation.

For the record, the flat representation of an anonymous struct with a
an integer and a char data members is the string:

            'struct {int i; char c;}'

	* src/abg-ir.cc ({class, union}_decl::get_pretty_representation):
	Use the flat representation of the class or union even for
	internal purposes.
	* tests/data/test-annotate/libtest23.so.abi: Adjust.
	* tests/data/test-diff-dwarf/PR25058-liblttng-ctl-report-1.txt: Likewise.
	* tests/data/test-diff-filter/test30-pr18904-rvalueref-report0.txt: Likewise.
	* tests/data/test-diff-filter/test30-pr18904-rvalueref-report1.txt: Likewise.
	* tests/data/test-diff-filter/test30-pr18904-rvalueref-report2.txt: Likewise.
	* tests/data/test-diff-filter/test35-pr18754-no-added-syms-report-0.txt: Likewise.
	* tests/data/test-diff-pkg/spice-server-0.12.4-19.el7.x86_64-0.12.8-1.el7.x86_64-report-2.txt: Likewise.
	* tests/data/test-read-dwarf/PR25007-sdhci.ko.abi: Likewise.
	* tests/data/test-read-dwarf/libtest23.so.abi: Likewise.
	* tests/data/test-read-dwarf/test11-pr18828.so.abi: Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>