writer: Avoid using dynamic hashing in type maps
When using type maps the hash used for a given type T is usually the
pointer value of the canonical type of T.

In the rare cases where T doesn't have a canonical type, we were using
a 'dynamic hash' computed by recursively walking the components of T
and progressively building a hash that way.

The dynamic hashing code hasn't been updated in a while and would need
some overhaul to support hashing with schemes like MD5 or maybe sha
even.  It might be useful for various use cases that have been
proposed by some users over the years but nobody was motivated enough
to implement it.

In the mean time, rather than trying to come up with a fully beefed up
dynamic hashing code, we'd rather just return a constant number for
non canonicalized types.  In practise that amounts to forcing the code
of the maps to always use structural comparison for those non
canonicalized types.

Note that the amount of non-canonicalized types should be fairly
small.  For now, the only non-canonicalized types should be
declaration-only types and those are quite fast to compare anyway.

This patch thus introduces a new hashing scheme for maps in the
writer which just uses a numerical constant as the hash for
non-canonicalized types.

	* include/abg-fwd.h (hash_as_canonical_type_or_constant): Declare ...
	* src/abg-ir.cc (hash_as_canonical_type_or_constant): ... new
	function.
	* src/abg-writer.cc (type_hasher::operator()): Use the new
	hash_as_canonical_type_or_constant.
	* tests/data/test-read-dwarf/test16-pr18904.so.abi: Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>